+++ android/app/src/main/java/com/xflip/MainActivity.kt
@@ -1,4 +1,4 @@
-package id.belink
+package com.xflip
 
 import android.os.Bundle
 import com.facebook.react.ReactActivity
@@ -16,7 +16,7 @@ class MainActivity : ReactActivity() {
      * Returns the name of the main component registered from JavaScript. This is used to schedule
      * rendering of the component.
      */
-    override fun getMainComponentName(): String = "Belink"
+    override fun getMainComponentName(): String = "xFlip"
 
     /**
      * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]
