+++ android/app/src/main/java/com/xflip/MainApplication.kt
@@ -1,4 +1,4 @@
-package id.belink
+package com.xflip
 
 import android.app.Application
 import com.facebook.react.PackageList
