+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "Belink",
-  "displayName": "Belink"
+  "name": "xFlip",
+  "displayName": "xFlip"
 }
