+++ ios/Podfile
@@ -14,7 +14,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'Belink' do
+target 'xFlip' do
   config = use_native_modules!
 
   use_react_native!(
@@ -23,7 +23,7 @@ target 'Belink' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'BelinkTests' do
+  target 'xFlipTests' do
     inherit! :complete
     # Pods for testing
   end
