+++ ios/xFlip.xcodeproj/xcshareddata/xcschemes/xFlip.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "Belink.app"
-               BlueprintName = "Belink"
-               ReferencedContainer = "container:Belink.xcodeproj">
+               BuildableName = "xFlip.app"
+               BlueprintName = "xFlip"
+               ReferencedContainer = "container:xFlip.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "BelinkTests.xctest"
-               BlueprintName = "BelinkTests"
-               ReferencedContainer = "container:Belink.xcodeproj">
+               BuildableName = "xFlipTests.xctest"
+               BlueprintName = "xFlipTests"
+               ReferencedContainer = "container:xFlip.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Belink.app"
-            BlueprintName = "Belink"
-            ReferencedContainer = "container:Belink.xcodeproj">
+            BuildableName = "xFlip.app"
+            BlueprintName = "xFlip"
+            ReferencedContainer = "container:xFlip.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Belink.app"
-            BlueprintName = "Belink"
-            ReferencedContainer = "container:Belink.xcodeproj">
+            BuildableName = "xFlip.app"
+            BlueprintName = "xFlip"
+            ReferencedContainer = "container:xFlip.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
