+++ ios/xFlip/Images.xcassets/AppIcon.appiconset/Contents.json
