+++ ios/xFlipTests/xFlipTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface BelinkTests : XCTestCase
+@interface xFlipTests : XCTestCase
 
 @end
 
-@implementation BelinkTests
+@implementation xFlipTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
