+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "belink",
+  "name": "xflip",
   "version": "0.0.1",
   "private": true,
   "scripts": {
